Monkey_Locker_Cervesa/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlmodel
sqlalchemy
psycopg2-binary
deepface
opencv-python
dlib
mediapipe
numpy
pandas
matplotlib
scipy
tensorflow
python-jose[cryptography]
passlib[bcrypt]
httpx
pydantic
boto3
python-dotenv
botocore